requirements.txt
pytextspan~=0.5.0
numpy>=1.12
jdatetime>=4.1.0
python-dateutil>=2.8.2
setuptools>=56.0.0